Move rendering backend impls to framework/render
