+++ src/framework/backend_null.rs
@@ -13,6 +13,7 @@ use crate::framework::backend::{
 use crate::framework::context::Context;
 use crate::framework::error::GameResult;
 use crate::framework::graphics::BlendMode;
+use crate::framework::render::null_impl::NullRenderer;
 use crate::game::Game;
 
 pub struct NullBackend;
@@ -57,91 +58,6 @@ impl BackendEventLoop for NullEventLoop {
         ctx.imgui.borrow_mut().fonts().build_alpha8_texture();
 
         Ok(Box::new(NullRenderer))
-    }
-
-    fn as_any(&self) -> &dyn Any {
-        self
-    }
-}
-
-pub struct NullTexture(u16, u16);
-
-impl BackendTexture for NullTexture {
-    fn dimensions(&self) -> (u16, u16) {
-        (self.0, self.1)
-    }
-
-    fn add(&mut self, _command: SpriteBatchCommand) {}
-
-    fn clear(&mut self) {}
-
-    fn draw(&mut self) -> GameResult<()> {
-        Ok(())
-    }
-
-    fn as_any(&self) -> &dyn Any {
-        self
-    }
-}
-
-pub struct NullRenderer;
-
-impl BackendRenderer for NullRenderer {
-    fn renderer_name(&self) -> String {
-        "Null".to_owned()
-    }
-
-    fn clear(&mut self, _color: Color) {}
-
-    fn present(&mut self) -> GameResult {
-        Ok(())
-    }
-
-    fn create_texture_mutable(&mut self, width: u16, height: u16) -> GameResult<Box<dyn BackendTexture>> {
-        Ok(Box::new(NullTexture(width, height)))
-    }
-
-    fn create_texture(&mut self, width: u16, height: u16, _data: &[u8]) -> GameResult<Box<dyn BackendTexture>> {
-        Ok(Box::new(NullTexture(width, height)))
-    }
-
-    fn set_blend_mode(&mut self, _blend: BlendMode) -> GameResult {
-        Ok(())
-    }
-
-    fn set_render_target(&mut self, _texture: Option<&Box<dyn BackendTexture>>) -> GameResult {
-        Ok(())
-    }
-
-    fn draw_rect(&mut self, _rect: Rect<isize>, _color: Color) -> GameResult {
-        Ok(())
-    }
-
-    fn draw_outline_rect(&mut self, _rect: Rect<isize>, _line_width: usize, _color: Color) -> GameResult {
-        Ok(())
-    }
-
-    fn set_clip_rect(&mut self, _rect: Option<Rect>) -> GameResult {
-        Ok(())
-    }
-
-    fn draw_triangles(
-        &mut self,
-        _vertices: &[VertexData],
-        _texture: Option<&Box<dyn BackendTexture>>,
-        _shader: BackendShader,
-    ) -> GameResult<()> {
-        Ok(())
-    }
-
-    fn draw_triangles_indexed(
-        &mut self,
-        vertices: &[VertexData],
-        indices: super::graphics::IndexData,
-        texture: Option<&Box<dyn BackendTexture>>,
-        shader: BackendShader,
-    ) -> GameResult {
-        Ok(())
     }
 
     fn as_any(&self) -> &dyn Any {
