+++ src/framework/mod.rs
@@ -14,8 +14,7 @@ pub mod filesystem;
 pub mod gamepad;
 pub mod graphics;
 pub mod keyboard;
-#[cfg(feature = "render-opengl")]
-pub mod render_opengl;
+pub mod render;
 pub mod ui;
 pub mod util;
 pub mod vfs;
